move change password modal to its own file
